requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-highlight==1.2.0
Django==2.2
django-rest-swagger==2.2.0
djangorestframework==3.10.2
drf-yasg==1.16.1
entrypoints==0.3
flake8==3.7.8
idna==2.8
importlib-metadata==0.19
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
openapi-codec==1.3.2
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.1.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
